refactor numeric_factorization and readability improvements
