+++ include/openmc/bateman_solvers.h
@@ -12,6 +12,34 @@
 namespace openmc {
 
 //==============================================================================
+//! Numeric LU factorization values for a CRAM shifted linear system
+//!
+//! Stores the complex-valued L/U entries for (dt*A - theta*I) corresponding
+//! to a previously computed SymbolicLUFactorization. The U diagonal is stored
+//! both in the CSC values array and as precomputed reciprocals for faster back
+//! substitution.
+//==============================================================================
+
+struct NumericLUFactorization {
+  vector<std::complex<double>> l_data;
+  vector<std::complex<double>> u_data;
+  vector<std::complex<double>> u_diag_inv;
+};
+
+//! Numeric left-looking LU factorization of the CRAM shifted operator
+//! (dt*A - theta*I) without explicitly forming it. Uses no pivoting; this is
+//! safe because A is Metzler (non-negative off-diagonal) and |Im(theta)| >=
+//! 1.194 for every CRAM pole, so every pivot satisfies |u_jj| >= |Im(theta)|.
+void numeric_factorize_cram(const CSCMatrix& A, double dt,
+  std::complex<double> theta, const SymbolicLUFactorization& symbolic,
+  NumericLUFactorization& numeric, vector<std::complex<double>>& work);
+
+//! Solve LUx = b using a symbolic LU pattern and matching numeric values.
+void triangular_solve_lu(const vector<double>& b,
+  const SymbolicLUFactorization& symbolic,
+  const NumericLUFactorization& numeric, vector<std::complex<double>>& x);
+
+//==============================================================================
 //! Abstract base class for Bateman equation solvers
 //!
 //! Solves dN/dt = A*N over an interval dt, given initial composition N(0).
@@ -64,11 +92,11 @@ public:
     int substeps = 1) override;
 
 private:
-  // --- CRAM coefficients ---
-  int n_poles_;                        //!< Number of poles (k/2)
-  vector<std::complex<double>> alpha_; //!< Residues [n_poles]
-  vector<std::complex<double>> theta_; //!< Poles [n_poles]
-  double alpha0_;                      //!< Limit at infinity
+  // --- CRAM coefficients (non-owning views of the static tables) ---
+  int n_poles_;                               //!< Number of poles (k/2)
+  const std::complex<double>* alpha_;         //!< Residues [n_poles]
+  const std::complex<double>* theta_;         //!< Poles [n_poles]
+  double alpha0_;                             //!< Limit at infinity
 };
 
 } // namespace openmc
